Significantly refactored code and repository. All functions to compute solar strap heritability are now in h2o_utility.py. Other support packages and familial recurrence have been removed.
Added the ability to compute both AE and ACE models with the 'ace=yes' option. The default
is to only compute the AE model. If ace=yes is provided on the command line then the ACE model
will also be computed using the house option in solar. The mother id is used as the hhid.

Added the ability to toggle the verbosity of the output with the 'verbose=yes' option.

Rhinitis example still works the same as before.